refactor: Controller 로직 분리 및 서비스 계층 적용
