+++ src/main/java/community/whatever/onembackendjava/urlshorten/UrlShortenController.java
@@ -1,0 +1,23 @@
+package community.whatever.onembackendjava.urlshorten;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class UrlShortenController {
+
+    private final UrlShortenService urlShortenService;
+
+    @PostMapping("/shorten-url/search")
+    public String shortenUrlSearch(@RequestBody String key) {
+        return urlShortenService.shortenUrlSearch(key);
+    }
+
+    @PostMapping("/shorten-url/create")
+    public String shortenUrlCreate(@RequestBody String originUrl) {
+        return urlShortenService.shortenUrlCreate(originUrl);
+    }
+}
