+++ src/main/java/community/whatever/onembackendjava/urlshorten/UrlShortenService.java
@@ -1,0 +1,24 @@
+package community.whatever.onembackendjava.urlshorten;
+
+import java.util.Random;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class UrlShortenService {
+
+    private final UrlShortenRepository urlShortenRepository;
+
+    public String shortenUrlSearch(String key) {
+        return urlShortenRepository.findByKey(key)
+            .orElseThrow(() -> new IllegalArgumentException("Invalid key"));
+    }
+
+    public String shortenUrlCreate(String originUrl) {
+        String randomKey = String.valueOf(new Random().nextInt(10000));
+        urlShortenRepository.save(randomKey, originUrl);
+        return randomKey;
+    }
+
+}
